requirements.txt
discord.py
httpx
mailslurp_client
jishaku